Squashed commit of the following (senders fix and colors):
(1) When loading previous messages on startup, the sender is now nullified for messages of the current user (as in the normal flow also when adding a new message, allowing the frontend to display messages of the current user in one side (left) and the others in the right side).
(2) Added an algorithm to select a different color for every sender.