Squashed commit of the following:
(1) WebSocketService.tsx: added a text message for the connections during $connect handling.
(2) To be researched further:
              // connections: [],
              // lastIncomingMessageTimestamp: action.payload ? new Date().toLocaleString('en-GB', options) : '',
(3) ScheduleExpression: cron(0/5 8-22 * * ? *) # Every 5 minutes
------------------------------------------------
ConnectionsUsernamesLayer:
    ..
          Description: Layer for collecting an array [{connection, username}].
------------------------------------------------
Follow-up on the previous commit:
    export interface WebsocketsState {
      isConnected: boolean;
      connections: IConnection[];
      lastIncomingMessageTimestamp: string;
    }

Refactored $connect and sendRandomMessage to send an array of connections and usernames separate from the messages, and saved it in the frontend, to be displayed as a tooltip and onClick the network icon.

(1) const EXPIRATION_TIME = 12 * 60 * 60; // --> 12 hours
(2) ScheduleExpression: cron(* 8-22 * * ? *) # --> Every minute from 08:00 to 23:00
    if (new Date().getMinutes() === 0)
            content = record.content;
         else
            content = prepareConnectedUsersMessage