Squashed commit of the following (connections management):
When the client becomes visible (from invisible), the client now pings the server and waits for few seconds. If the server doesn't respond, the client reconnects.

Added a <div className='network-container'> for the list of connected users instead of displaying in the messages.

$disconnect handler: collectConnectionsAndUsernames

Housekeeping in the lambda functions: const AAA = process.env.AAA;

Invalidate the messages cache when a record is modified or deleted.