Squashed commit of the following (messages deletion):
Introduced support for deleting messages.
Senders can only delete their own messages.
Relocated the setting of message id with uuid from the backend to the frontend, to allow the frontend to delete a message that was just added by the current user.

Fixed a mistake from one of the previous commits: The $connect handler sent by mistake all connected usernames and the previous chat messages to all connected usernames.
The right thing to do is to send all connected usernames (including the current new one) to all connected users, but send the previous chat messages only to the current client.

Housekeeping in $disconnect handler: currentUserId was retrieved for no reason.

Renamed the SendMessage route key to GenericWebsocketReceiver, to also handle delete notifications in addition to previous responsiblities.